refactor: split internal.rs into internal/ directory — F1
Mechanical split of 1,356-line file into 4 files per dependable-rust:
- internal/mod.rs — wasm_engine() singleton, PluginManifest, re-exports
- internal/mother_child.rs — bindings, PluginEngine, WasmChild adapter
- internal/command.rs — command_bindings, CommandHostState, CommandEngine
- internal/tests.rs — all 22 tests with shared fixtures

Zero logic changes. All 22 plugin tests pass unchanged.
plugin/mod.rs stays identical (same 12-line interface).
